Update input method
Now the inputs for Gaudi of DSTs from the grid are configured automatically via Dirac commands, which have been added to the work flow for automation.

Following that, the old lists of PFNs for the different samples have been deleted.

This allows for further modularization of the scripts, since now the input list is isolated. Now no changes to the scripts must be made if one wants to use ganga.

The number of events have been reduced to 20k/sample, as there are samples just too big to run on lxplus